requirements.txt
pytube
openai-whisper
